makesymlinks.sh shouldn't try to move files that don't exist

+++ makesymlinks.sh
@@ -27,8 +27,10 @@ echo "...done"
 # move any existing dot_files in homedir to dot_files_old directory, then create symlinks
 for file in $files; do
     echo "Found $file"
-    echo "Moving any existing dot_files from ~ to $olddir"
-    mv ~/$file ~/dot_files_old/
+    if [ -f ~/$file || -d ~/$file ]; then
+        echo "Moving any existing dot_files from ~ to $olddir"
+        mv ~/$file ~/dot_files_old/
+    fi
     echo "Creating symlink to $file in home directory."
     ln -s $dir/$file ~/$file
 done
